requirements.txt
asgiref==3.9.1
Django==5.2.6
gunicorn==23.0.0
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.9.0
djangorestframework==3.20.1